services/static-web-challenges/requirements.txt
pyyaml==6.0.1
pyjwt==2.8.0
